refactor: 배송 조회 로직 리팩터링 및 컴파일 에러 해결
